test for symlink before unlinking

+++ install.sh
@@ -4,7 +4,7 @@ DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 for f in .*; do
     if [[ ! $f =~ ^\.(git|\.)?$ ]]; then
-        unlink "$HOME/$f"
+        [ -h "$HOME/$f" ] && unlink "$HOME/$f"
         ln -s "$DIR/$f" "$HOME/$f"
     fi
 done
